Docs Move Style component; spellcheck comments
