+++ docs/victory-pie/docs.js
@@ -1,25 +1,24 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import Ecology from "ecology";
-import Radium, { Style, StyleRoot } from "radium";
+import Radium from "radium";
 import * as docgen from "react-docgen";
 import { merge, random, range } from "lodash";
-import { VictoryPie } from "../src/index";
-import { VictoryTheme, appendLinkIcon, ecologyPlaygroundLoading } from "formidable-landers";
+import { VictoryPie } from "../../src/index";
+import { appendLinkIcon, ecologyPlaygroundLoading } from "formidable-landers";
 
 class Docs extends React.Component {
   render() {
     return (
-      <StyleRoot>
+      <div>
         <Ecology
           overview={require("!!raw!./ecology.md")}
-          source={docgen.parse(require("!!raw!../src/components/victory-pie"))}
+          source={docgen.parse(require("!!raw!../../src/components/victory-pie"))}
           scope={{merge, random, range, React, ReactDOM, VictoryPie}}
           playgroundtheme="elegant"
           customRenderers={merge(appendLinkIcon, ecologyPlaygroundLoading)}
         />
-        <Style rules={VictoryTheme}/>
-      </StyleRoot>
+      </div>
     );
   }
 }
