+++ package.json
@@ -45,6 +45,7 @@
     "react": "~15.1.0",
     "react-addons-test-utils": "~15.1.0",
     "react-dom": "~15.1.0",
+    "react-router": "^2.4.0",
     "sinon": "^1.15.4",
     "sinon-chai": "^2.8.0"
   }
